Reminders to Record Your Work

It's easy to get carried away and forget to record your work. stow can send you a reminder if you have work that needs to be committed, but haven't used a stow function for some time.

remind_me(after = 60)

Arguments

after
how long (in minutes) to wait before reminding you to record your work

Details

To turn off reminders, just do remind_me(after = 0). remind_me() is called by default when you run create_repo().